# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediERP - Medical Store ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: isActive
          in: query
          schema: { type: boolean }
        - name: lowStock
          in: query
          schema: { type: boolean }
        - name: nearExpiry
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Deactivate a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}/clear-stock:
    post:
      operationId: clearProductStock
      tags: [products]
      summary: Clear stock without journal entries
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Stock cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  # ── INVENTORY ─────────────────────────────────────────────
...
  /journals:
    get:
      operationId: listJournals
      tags: [journals]
      summary: List journal entries
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: accountId
          in: query
          schema: { type: integer }
        - name: type
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournal
      tags: [journals]
      summary: Create a manual journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /journals/{id}:
    get:
      operationId: getJournal
      tags: [journals]
      summary: Get a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    delete:
      operationId: deleteJournal
      tags: [journals]
      summary: Delete a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"

  # ── VENDORS ───────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, sellingPricePerUnit, reorderLevel, isActive, totalTablets, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        genericName: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        unitType:
          type: string
        tabsPerPack: { type: integer }
        packsPerBox: { type: integer }
        sellingPricePerUnit: { type: number }
        sellingPricePerPack: { type: number }
        sellingPricePerBox: { type: number }
        reorderLevel: { type: integer }
        isActive: { type: boolean }
        totalTablets: { type: integer }
        totalPacks: { type: number }
        totalBoxes: { type: number }
        nearestExpiry: { type: ["string", "null"] }
        createdAt: { type: string }
    ProductInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        genericName: { type: string }
        category: { type: string }
        reorderLevel: { type: integer }
    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        genericName: { type: string }
        category: { type: string }
        reorderLevel: { type: integer }
        isActive: { type: boolean }

    # ── INVENTORY ────────────────────────────────────────────
...
    JournalEntry:
      type: object
      required: [id, date, description, type, lines, createdAt]
      properties:
        id: { type: integer }
        date: { type: string, format: date }
        description: { type: string }
        reference: { type: ["string", "null"] }
        type:
          type: string
          enum: [manual, sale, purchase, expense, adjustment]
        lines:
          type: array
          items: { $ref: "#/components/schemas/JournalLine" }
        totalDebit: { type: number }
        totalCredit: { type: number }
        createdAt: { type: string }
    JournalLine:
      type: object
      required: [id, accountId, accountName, debit, credit]
      properties:
        id: { type: integer }
        accountId: { type: integer }
        accountName: { type: string }
        debit: { type: number }
        credit: { type: number }
        description: { type: ["string", "null"] }
    JournalInput:
      type: object
      required: [date, description, lines]
      properties:
        date: { type: string, format: date }
        description: { type: string }
        reference: { type: string }
        lines:
          type: array
          items: { $ref: "#/components/schemas/JournalLineInput" }
    JournalLineInput:
      type: object
      required: [accountId, debit, credit]
      properties:
        accountId: { type: integer }
        debit: { type: number }
        credit: { type: number }
        description: { type: string }

    # ── VENDOR ───────────────────────────────────────────────
...