# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediERP - Medical Store ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  genericName?: string | null;
  /** @nullable */
  category?: string | null;
  unitType?: string;
  tabsPerPack?: number;
  packsPerBox?: number;
  sellingPricePerUnit: number;
  sellingPricePerPack?: number;
  sellingPricePerBox?: number;
  reorderLevel: number;
  isActive: boolean;
  totalTablets: number;
  totalPacks?: number;
  totalBoxes?: number;
  /** @nullable */
  nearestExpiry?: string | null;
  createdAt: string;
}